requirements.txt
langchain
langchain-community
langchain-ollama
langgraph
requests
chainlit
pydantic
mcp
langchain-mcp-adapters
langchain-google-genai


ddgs
python-dotenv
langsmith
